# dispatcher/backend/docs/openapi_v1.yaml
openapi: 3.0.0
servers:
  - description: production server
    url: https://api.farm.openzim.org/v1
info:
  description: |
    Zimfarm is openZIM's automated & distributed ZIM files creation engine.
    openZIM periodically publishes updated ZIM files for a large variety of content.

    This API provides public access to the scheduled ZIM files to create as well as tracking of its tasks.

    This API is also used by _workers_ to fetch, process and update those tasks.
  version: 1.0.0
  title: Zimfarm API
  contact:
    email: contact+zimfarm@kiwix.org
    url: https://kiwix.org
  license:
    name: GNU General Public License v3.0
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
...
paths:
...
  /tasks/:
    get:
      tags:
      - public
      summary: list tasks
      operationId: listTasks
      description: Query the Tasks database
      parameters:
      - $ref: '#/components/parameters/SkipParameter'
      - $ref: '#/components/parameters/LimitParameter'
      - $ref: '#/components/parameters/StatusParameter'
      - $ref: '#/components/parameters/ScheduleNameQueryParameter'
      responses:
        200:
          description: List of Tasks matching criteria
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: '#/components/schemas/MetaProperty'
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/TaskItem'
        400:
          description: Bad Request (invalid input)
  /tasks/{taskId}:
    get:
      tags:
      - public
      summary: get task detail
      operationId: getTask
      description: Retrieve a Task's details
      parameters:
      - $ref: '#/components/parameters/TaskIdParameter'
      responses:
        200:
          description: Task record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        400:
          description: Bad Request (invalid input)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InputError'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
      - workers
      summary: create task
      operationId: addTask
      description: Turn a RequestedTask into a Task
      security:
        - token: []
        - oauth: []
      parameters:
      - $ref: '#/components/parameters/TaskIdParameter'
      responses:
        201:
          description: Task Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    $ref: '#/components/schemas/MongoId'
        400:
          description: Bad Request (invalid input)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InputError'
        401:
          description: Unauthorized
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        423:
          description: "Unable to create Task: created by another worker"
    patch:
      tags:
      - workers
      summary: update task
      operationId: editTask
      description: Update a Task's details
      security:
        - token: []
        - oauth: []
      parameters:
      - $ref: '#/components/parameters/TaskIdParameter'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - event
                - payload
              properties:
                event:
                  $ref: '#/components/schemas/TaskStatus'
                payload:
                  type: object
                  additionalProperties: true
                  description: event-related payload
                  example:
                    task_log: agbcdls213821983hjhkj-json.log
      responses:
        204:
          description: Task Updated
        400:
          description: Bad Request (invalid input)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InputError'
        401:
          description: Unauthorized
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tasks/{taskId}/cancel:
    post:
      tags:
      - restricted
      summary: cancel task
      operationId: cancelTask
      description: Request cancellation of a Task. Actual cancellation is asyncrhonous.
      security:
        - token: []
        - oauth: []
      parameters:
      - $ref: '#/components/parameters/TaskIdParameter'
      responses:
        204:
          description: Cancelation Requested
        401:
          description: Unauthorized
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /users/{username}/keys/:
    get:
      tags:
      - restricted
      summary: get user's SSH keys
      operationId: getUserKeys
      description: Retrieve the list of registered SSH keys for User
      security:
        - token: []
        - oauth: []
      parameters:
      - $ref: '#/components/parameters/UsernameParameter'
      responses:
        200:
          description: List of SSH Keys for User
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required:
                    - added
                    - fingerprint
                    - key
                    - last_used
                    - name
                    - pkcs8_key
                    - type
                  properties:
                    added:
                      type: string
                      format: date-time
                      example: "2019-05-15T09:43:37.216000Z"
                    fingerprint:
                      type: string
                      example: c470c5ba0110ce26fec082d233930fad
                    key:
                      type: string
                      example: AAAAB3NzaC1yc2EAAAADAQABAAABAQDDIRSam+GcKAAnm0jF\/VB12rtXCzpFE1gsT9CnFw5dYbP
                    last_used:
                      type: string
                      format: date-time
                      example: "2019-08-12T08:22:10.519000Z"
                    name:
                      type: string
                      example: my-key
                    pkcs8_key:
                      type: string
                      example: -----BEGIN PUBLIC KEY-----\nMIIBIjANBgkqhk
                    type:
                      type: string
                      enum:
                        - RSA
                      example: RSA
        401:
          description: Unauthorized
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
      - restricted
      summary: add SSH Key
      operationId: addUserKeys
      description: Upload a new SSH Public Key for User
      security:
        - token: []
        - oauth: []
      parameters:
      - $ref: '#/components/parameters/UsernameParameter'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - name
                - key
              properties:
                name:
                  type: string
                  example: my-key
                  minLength: 1
                key:
                  type: string
                  description: Content part of an RSA _Public_ Key
                  minLength: 1
                  example: AAAAB3NzaC1yc2EAAAADAQABAAABAQDDIRSam+GcKAAnm0jF\/VB12rtXCzpFE1gsT9CnFw5dYbP
      responses:
        201:
          description: SSH Key Uploaded
        400:
          description: Bad Request (invalid input)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InputError'
        401:
          description: Unauthorized
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{username}/keys/{fingerprint}:
    delete:
      tags:
      - restricted
      summary: delete key
      operationId: deleteUserKey
      description: Remove SSH Key from User
      security:
        - token: []
        - oauth: []
      parameters:
      - $ref: '#/components/parameters/UsernameParameter'
      - in: path
        required: true
        name: fingerprint
        schema:
          type: string
          description: SSH fingerprint of the key
      responses:
        204:
          description: SSH Key Deleted
        401:
          description: Unauthorized
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    token:
      type: apiKey
      description: API Token
      name: token
      in: header
    oauth:
      type: oauth2
      description: OAuth2
      flows:
        password:
          tokenUrl: https://api.farm.openzim.org/v1/auth/oauth2
          refreshUrl: https://api.farm.openzim.org/v1/auth/oauth2
  parameters:
    SkipParameter:
      in: query
      name: skip
      description: number of records to skip for pagination (yay)
      required: false
      schema:
        $ref: '#/components/schemas/SkipProperty'
    LimitParameter:
      in: query
      name: limit
      description: maximum number of records to return
      required: false
      schema:
        $ref: '#/components/schemas/LimitProperty'
...
    ScheduleNameQueryParameter:
      in: query
      required: false
      name: schedule_name
      description: filter by schedule_name
      schema:
        $ref: '#/components/schemas/ScheduleName'
    UsernameParameter:
      in: path
      required: true
      name: username
      schema:
        $ref: '#/components/schemas/Username'
    TaskIdParameter:
      in: path
      name: taskId
      required: true
      schema:
        type: string
    StatusParameter:
      in: query
      name: status
      description: only those matching this statuses
      required: false
      schema:
        $ref: '#/components/schemas/TaskStatus'
  schemas:
    Language:
      type: object
      required:
      - code
      - name_en
      - name_native
      properties:
        code:
          type: string
          format: ISO 639-1 language code
          example: fr
        name_en:
          type: string
          example: French
        name_native:
          type: string
          example: Français
...
    ScheduleCategory:
      type: string
      example: wikipedia
      enum:
        - gutenberg
        - other
        - phet
        - psiram
        - stack_exchange
        - ted
        - vikidia
        - wikibooks
        - wikinews
        - wikipedia
        - wikiquote
        - wikisource
        - wikispecies
        - wikiversity
        - wikivoyage
        - wiktionary
    Schedule:
      type: object
      additionalProperties: false
      required:
        - category
        - config
        - enabled
        - language
        - name
        - tags
      properties:
        category:
          $ref: '#/components/schemas/ScheduleCategory'
        config:
          $ref: '#/components/schemas/Config'
        enabled:
          type: boolean
          example: true
        language:
          $ref: '#/components/schemas/Language'
        name:
          $ref: '#/components/schemas/ScheduleName'
        most_recent_task:
          type: object
          required:
            - _id
            - status
            - updated_at
          properties:
            _id:
              $ref: '#/components/schemas/MongoId'
            status:
              $ref: '#/components/schemas/TaskStatus'
            updated_at:
              type: string
              format: date-time
        tags:
          $ref: '#/components/schemas/ScheduleTags'
    ScheduleName:
      type: string
      example: wikipedia_fr_all
    Config:
      type: object
      required:
        - flags
        - image
        - resources
        - task_name
        - warehouse_path
      properties:
        flags:
          type: object
        image:
          $ref: '#/components/schemas/ConfigImage'
        resources:
          $ref: '#/components/schemas/ConfigResources'
        task_name:
          type: string
          example: mwoffliner
        warehouse_path:
          type: string
          example: /wikipedia
    ConfigImage:
      type: object
      required:
        - name
        - tag
      properties:
        name:
          type: string
          example: openzim/mwoffliner
        tag:
          type: string
          example: latest
    ConfigResources:
      type: object
      required:
        - cpu
        - memory
        - disk
      properties:
        cpu:
          $ref: '#/components/schemas/Cpu'
        memory:
          $ref: '#/components/schemas/Memory'
        disk:
          $ref: '#/components/schemas/Disk'
    Cpu:
      type: integer
      format: int32
      example: 3
      description: number of CPU cores required to run it
    Memory:
      type: integer
      format: int32
      example: 1024
      description: RAM (in bytes) required to run it
    Disk:
      type: integer
      format: int32
      description: Disk space (in bytes) required to run it
...
    TaskItem:
      type: object
      required:
        - _id
        - status
        - timestamp
      properties:
        _id:
          $ref: '#/components/schemas/MongoId'
        status:
          $ref: '#/components/schemas/TaskStatus'
        timestamp:
          $ref: '#/components/schemas/TaskTimestamp'
        worker:
          $ref: '#/components/schemas/WorkerName'
        schedule_name:
          $ref: '#/components/schemas/ScheduleName'
...
    Task:
      type: object
      required:
        - _id
        - config
        - status
        - events
        - schedule
        - schedule_name
        - timestamp
        - worker
      properties:
        _id:
          $ref: '#/components/schemas/MongoId'
        config:
          $ref: '#/components/schemas/Config'
        status:
          $ref: '#/components/schemas/TaskStatus'
        events:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        schedule:
          $ref: '#/components/schemas/Schedule'
        schedule_name:
          $ref: '#/components/schemas/ScheduleName'
        timestamp:
          $ref: '#/components/schemas/TaskTimestamp'
        worker:
          $ref: '#/components/schemas/WorkerName'
        container:
          type: object
          properties:
            command:
              type: array
              items:
                type: string
              example:
                - mwoffliner
                - mwUrl
                - http://fr.wikipedia.org
            exit_code:
              type: integer
              format: int32
              example: 1
            image:
              type: string
              example: openzim/mwoffliner:latest
            log:
              type: string
              example: "das878sd7sa87f8f78dsf7dsds.log"
        debug:
          type: object
          properties:
            exception:
              type: string
              example: "SyntaxError at line 1"
            traceback:
              type: string
              example: null
            log:
              type: string
              example: "[2019-11-18 15:49:30,337: INFO] starting zimfarm task-worker"
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Not Found
    InputError:
      type: object
      properties:
        message:
          oneOf:
            - type: object
              example: {"schedule_name": "required field"}
            - type: string
              example: Invalid signature format (not base64)
...
    MetaProperty:
      type: object
      properties:
        count:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        skip:
          type: integer
          example: 0
    LimitProperty:
      type: integer
      format: int32
      minimum: 0
      example: 10
      default: 10
    SkipProperty:
      type: integer
      format: int32
      minimum: 0
      example: 0
      default: 0
    MongoId:
      type: string
      format: uuid
      example: 5dbc599b3b61e21c362703e4
    Username:
      type: string
      example: john
      minLength: 1
...
    ScheduleTags:
      type: array
      items:
        type: string
      example:
        - selection
        - full
        - top
    TaskStatus:
      type: string
      example: canceled
      enum:
        - started
        - scraper_started
        - scraper_completed
        - scraper_killed
        - failed
        - cancel_requested
        - canceled
        - succeeded
        - created_file
        - uploaded_file
    SingleTimestamp:
      type: string
      format: date-time
      example: "2019-11-19T15:36:06.563000Z"
    Event:
      type: object
      properties:
        code:
          $ref: '#/components/schemas/TaskStatus'
        timestamp:
          $ref: '#/components/schemas/SingleTimestamp'
    WorkerName:
      type: string
      example: worker-name
    TaskTimestamp:
      type: object
      required:
        - requested
      properties:
        requested:
          $ref: '#/components/schemas/SingleTimestamp'
        reserved:
          $ref: '#/components/schemas/SingleTimestamp'
        started:
          $ref: '#/components/schemas/SingleTimestamp'
        scraper_started:
          $ref: '#/components/schemas/SingleTimestamp'
        scraper_completed:
          $ref: '#/components/schemas/SingleTimestamp'
        scraper_killed:
          $ref: '#/components/schemas/SingleTimestamp'
        failed:
          $ref: '#/components/schemas/SingleTimestamp'
        cancel_requested:
          $ref: '#/components/schemas/SingleTimestamp'
        canceled:
          $ref: '#/components/schemas/SingleTimestamp'
        succeeded:
          $ref: '#/components/schemas/SingleTimestamp'
...